src/requirements.txt
asttokens==2.4.1
beautifulsoup4==4.12.3
colorama==0.4.6
coverage==7.6.0
decorator==5.1.1
docopt==0.6.2
executing==2.0.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
libsass==0.23.0
livereload==2.7.0
Markdown==3.6
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-watch==4.2.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
watchdog==4.0.1
wcwidth==0.2.13
